requirements.txt
langchain>=0.1.16
langchain-openai>=0.1.0
python-dotenv==1.0.0
flask
flask-cors
requests
gunicorn
